# Update query planner to allow multiple starting points