## What is the goal of this PR?

We implement a new formulation of the optimisation problem at the heart of query planning that allows the new traversal algorithm to start from multiple points in the query graph. The new query planner also takes into account the specifics of the traversal engine to directly optimise the expected computational costs of traversal.

## What are the changes implemented in this PR?

### Summary

The new vertex-ordering query planner places no restriction on the number of starting vertices to allow the new traversal engine to fully leverage the intersection iterators. The bulk of the meaningful change is contained in GraphPlanner and PlannerEdge/Vertex. This change leads to minor tweaks in GraphProcedure and GraphIterator that allow us to maintain compatibility between the classes.

The linear solver was also switched from the mixed-integer (that is, mixing integer and floating-point variables) SCIP mode to an integer SAT mode, which is optimised around all the variables and coefficients being integer valued, and as such allows the solver to arrive at a better or optimal solution faster.

### Query Planner

The new query planner attempts to directly estimate the number of database operations performed during traversal for a given query plan and minimizes that function. 

The traversal is performed as follows. For each value in the traversal procedure, the number of values we expect to hit every time we enter it is estimated to be either the minimum among the traversal cost of its incoming edges, or, if it has no incoming edges, the estimated size of the vertex itself. As the traversal is performed in an ordered fashion, the number of times we enter a vertex—and, therefore, the number of times we open an iterator into it and iterate through its values—depends straightforwardly on all the vertices we have visited before.

As an illustration, consider a diamond query graph with a given ordering.

```
   0
 ╱   ╲
1  —  2
 ╲   ╱
   3
```

Its traversal, in pseudocode, is performed like so:
```
i0 := iterate(0, {})
for v0 in i0:
    i1 := iterate(1, {v0})
    for v1 in i1:
        i2 := iterate(2, {v0, v1})
        for v2 in i2:
            i3 := iterate(3, {v1, v2})
            for v3 in i3:
                emit(v0, v1, v2, v3)
```

Let α(_v_) be the size of the vertex _v_, and β(_v_, _w_) be the number of values in _w_ that we expect to match for each value of the vertex _v_. Here is that same snippet, with each loop now annotated with the expected number of iterations:

```
i0 := iterate(0, {})
for v0 in i0:                           # α(0) iterations
    i1 := iterate(1, {v0})
    for v1 in i1:                       # β(0, 1) iterations
        i2 := iterate(2, {v0, v1})
        for v2 in i2:                   # min(β(0, 2), β(1, 2)) iterations
            i3 := iterate(3, {v1, v2})
            for v3 in i3:               # min(β(1, 3), β(2, 3)) iterations
                emit(v0, v1, v2, v3)
```

Note: we take minimum of βs, because the resulting set of values in a vertex is an intersection of the sets produced by all incoming edges, which has cardinality of at most min(β,..).

We will omit the parentheses in the α and β notation from this point on to improve legibility in the formulae.

It is easy to see that the number of iterators we create (or reset) in the previous snippet is

`seeks  = 1 + α0(1 + β01(1 + min(β02, β12)))`,

and the number of times we advance an iterator is 

`nexts  = α0(1 + β01(1 + min(β02, β12)(1 + min(β13, β23))))`.

Opening the brackets reveals that the difference between the two values is `α0 β01 min(β02, β12) min(β13, β23) - 1`. While seeks are significantly (5x to 10x) more expensive than nexts, optimising just them loses information on the cost of traversing the end vertices which will lead to the planner favouring expensive bipartite plans with as many ending vertices as possible. A simple solution therefore is to optimise the number of times the storage iterators are advanced. A more complex set of constraints taking that difference into account can be implemented in the future if needed.

To recap, so far the function we'd like to optimise looks like `α0(1 + β01(1 + min(β02, β12)(1 + min(β13, β23))))`. This is, however, not easily representable as a linear combination of some variables. The route we take is to instead minimize the function `(1 + α0) (1 + β01) (1 + min(β02, β12)) (1 + min(β13, β23))` (note the placement of parentheses). This formula strictly overestimates the number of operations performed.

This function generalizes to 

Π{_v_ ∈ S} (1 + α(_v_)) * Π{_v_ ∈ V}(1 + min({β(_x_, _v_): edge(_x_, _v_) is selected}))

where S is the set of starting vertices, V is the set of all vertices, and minimum of an empty set is defined to be 0.

Taking the natural logarithm of this expressions transforms the product into a sum of logarithms, which can be represented as a linear combination of boolean variables:

Σ{_v_ ∈ V} (log(1 + α(_v_)) * _v_:isStartingVertex) + Σ{_e_ ∈ E} (log(1 + β(_e_)) * _e_:isMinimal)

where:
* V is the set of all vertices,
* E is the set of all edges,
* β(_e_) is defined analogously to β(_v_, _w_) above (this accounts for edge multiplicity as well),
* _v_:isStartingVertex is true iff _v_ has no incoming edges,
* _e_:isMinimal is true iff _e_ is selected by the plan, and its cost is the lowest among all the incoming edges of its destination vertex.

This is the objective function of the new planner. The constraints serve to infer the values of these boolean variables from an ordering of the vertices.

### Performance measurements

The two commits we compare in this section are
* ec3b63054, which is exactly equivalent to master, but for the collection of the RocksDB usage statistics,
* a155d32b7, the latest commit in this PR at the time of measurement.

The main criterion we consider is the number of RocksDB operations, which we call here seeks and nexts. Timings are given as an intuitive reference only.

#### typedb-benchmark

The benchmark was run only once for each build. This was not intended as a true performance comparison, but as a smoke test, to make sure that the new planner does not degrage the server's performance.

With the build on current master, the simulation ran for 15 minutes and 45 seconds, performing 44 million seeks and 44 million nexts. 

With the new planner, the simulation ran for just over 12 minutes, performing 58 million seeks and 40 million nexts.

#### Complex queries (see #6550)

Each of the two queries—with and without attribute comparison—was run several times. After each query the statistics were collected and the server was restarted in order to purge all caches.

Both versions performed comparably on the query without attribute comparisons, performing around 3000 seeks and 2700 nexts and finishing within 600 to 700 ms.

On the query _with_ attribute comparisons, the version on master performed erratically, taking from 1 to 5 seconds to complete. It performed anywhere from 80'000 seeks, 300'000 nexts to 800'000 seeks, 600'000 nexts depending on the run. In contrast, the version in this PR performed around 3200 seeks and 10'000 next run after run, performance comparable to that of the query with no attribute comparisons.

### Smaller changes

* Improvements to `Optimiser`:
  * QoL improvements: native retrieval of the initial value, optimiser agnostic `value` retrieval;
  * allow to change constraints between optimiser passes;
  * bugfix: don't reset the model hint after first optimiser pass.
* Streamlined `GraphPlanner`:
  * easier to follow cost computation path:
    * costs computed separately from setting the objective function coefficients;
    * no back and forth calls between GraphPlanner and PlannerEdge/Vertex;
  * simplified value initialisation to a method;
  * methods reordered for readability.
* Stricter equality comparisons between `Identifier`s. 
* Stricter equality comparisons between `ProcedureEdge`s now that we can't rely on `order` for disambiguation.