@flyingsilverfin 2022-07-29
Update: we consider whether our cost function can also naturally be used to consider disconnected graphs.

```
0   2
|   |
1   3
```

In this mode, we don't have to do anything differently:
```
Π{v ∈ S} (1 + α(v)) * Π{v ∈ V}(1 + min({β(x, v): edge(x, v) is selected}))
```

the product of the `α(v)` already captures multiple possible starting points, and calculates cost according to our nested-loop formulation. The cost function however cannot capture the difference between executing the ordered plan: `0 - 1 - 2 - 3` versus `0 - 2 - 1 - 3`. We allow the traversal engine to perform the required short circuiting to make the two plans roughly equivalent in performance.